+++ app/api/core/infrastructure/facades/EntitiesFacade.ts
@@ -5,8 +5,6 @@ import { CreateEntityDTO, CreateEntitySchema } from '../express/entity/Schemas.j
 import { CreateEntityUseCaseFactory } from '../factories/CreateEntityUseCaseFactory.js';
 import { LoggerFactory } from '../factories/LoggerFactory.js';
 import { ExpressEntityMapper } from '../express/entity/ExpressEntityMapper.js';
-import { Entity } from '#api/core/domain/entity/Entity.js';
-import { CreateEntityFromPDFUseCaseFactory } from '../factories/CreateEntityFromPDFUseCaseFactory.js';
 
 export class EntityFacade {
   static async create(
@@ -23,17 +21,9 @@ export class EntityFacade {
 
       const input = ExpressEntityMapper.toEntityCreateInput({ dto: parsed, inputFiles });
 
-      let entity: Entity;
+      const useCase = CreateEntityUseCaseFactory.default({ targetLanguage });
 
-      if (Object.keys(parsed).length === 1 && parsed.title) {
-        const useCase = CreateEntityFromPDFUseCaseFactory.default({ targetLanguage });
-
-        entity = await useCase.execute(input);
-      } else {
-        const useCase = CreateEntityUseCaseFactory.default({ targetLanguage });
-
-        entity = await useCase.execute(input);
-      }
+      const entity = await useCase.execute(input);
 
       const duration = Date.now() - startTime;
