+++ app/api/core/infrastructure/factories/MongoEntitiesDAOFactory.ts
@@ -1,7 +1,7 @@
 import { ExecutionContext } from '#api/core/libs/ExecutionContext.js';
 import { User } from '#api/users.v2/model/User.js';
 import { getConnection } from '../mongodb/common/getConnectionForCurrentTenant.js';
-import { MongoEntitiesDAO } from '../mongodb/entity/MongoEntityDAO.js';
+import { MongoEntitiesDAO } from '../mongodb/entity/MongoEntitiesDAO.js';
 
 export class MongoEntitiesDAOFactory {
   static default(): MongoEntitiesDAO {
